Expose the UPDATE_PAYMENT_DETAILS service.
Before this patch, this payment app connected to the
UPDATE_PAYMENT_DETAILS service in Chrome for dynamic price updates.
However, not all implementations of user agents can expose a service.

This patch moves the implementation of UPDATE_PAYMENT_DETAILS service
into its own class SamplePaymentDetailsUpdateService that is exposed
from the payment app through AndroidManifest.xml, similar to
IS_READY_TO_PAY service. The PaymentActivity and
SamplePaymentDetailsUpdateService are connected through static member
variables in SamplePaymentDetailsUpdateService.

After this patch, user agents can connect to this payment app's
UPDATE_PAYMENT_DETAILS service for dynamic price updates.
