+++ SamplePay/app/src/main/aidl/org/chromium/components/payments/IPaymentDetailsUpdateServiceCallback.aidl
@@ -2,6 +2,8 @@ package org.chromium.components.payments;
 
 import android.os.Bundle;
 
+import org.chromium.components.payments.IPaymentDetailsUpdateService;
+
 /**
  * Helper interface used by the browser to notify the invoked native app about
  * merchant's response to one of the paymentmethodchange, shippingoptionchange,
@@ -22,4 +24,16 @@ interface IPaymentDetailsUpdateServiceCallback {
      * modified the payment details.
      */
     oneway void paymentDetailsNotUpdated();
+
+    /**
+     * Called during a payment flow to point the payment app back to the payment
+     * details update service to invoke when the user changes the payment
+     * method, the shipping address, or the shipping option.
+     *
+     * @param service The payment details update service to invoke when the user
+     *      changes the payment method, the shipping address, or the shipping
+     *      option.
+     */
+    oneway void setPaymentDetailsUpdateService(
+            IPaymentDetailsUpdateService service);
 }
